Extract backend test DB fake utilities into dedicated module
Move fake database and query matcher helpers from tests/conftest.py into
a dedicated test utility module at tests/test_utils/db_fakes.py. Helpers
moved: FakeInsertOneResult, FakeUpdateResult, FakeDeleteResult, FakeCursor,
FakeCollection, FakeDatabase, get_nested_values, matches_query, _set_nested,
_inc_nested, _eval_pipeline_expr.

Both tests/conftest.py and tests/api/conftest.py re-export the moved
symbols using the 'import X as X' idiom so all existing imports continue
to work without changes. FakeStorage remains in tests/conftest.py (it is
a storage fake, not a DB/query matcher helper).

Add 31 characterization tests for _eval_pipeline_expr covering all 9
issue-required cases: $eq, $ne, $cond (including truthy/falsy
non-boolean), $map, $filter, $$ variables, $ field references, plain
values, and nested expressions. Tests use string-literal assertions
independent of the function's location.

No product code changes. No behavior changes to query matching or
aggregation evaluation. No compatibility re-exports removed.

Closes #413
